+++ plugins/css-blank-pseudo/src/browser-global.js
@@ -1,3 +1,3 @@
 /* global self */
-import { default as cssBlankPseudo } from './browser';
-self.cssBlankPseudo = cssBlankPseudo;
+import { default as cssBlankPseudoInit } from './browser';
+self.cssBlankPseudoInit = cssBlankPseudoInit;