+++ plugins/css-blank-pseudo/.tape.mjs
@@ -17,11 +17,35 @@ postcssTape(plugin)({
 			preserve: false
 		}
 	},
-	'generated-selector-cases': {
-		message: 'correctly handles generated cases',
+	'basic:wrong-replacewith': {
+		message: 'correctly warns when replace with is invalid',
 		warnings: 1,
 		options: {
+			replaceWith: '#css-blank'
+		}
+	},
+	'examples/example': {
+		message: 'minimal example',
+	},
+	'examples/example:preserve-false': {
+		message: 'minimal example',
+		options: {
 			preserve: false
+		}
+	},
+	'examples/example:replacewith': {
+		message: 'minimal example',
+		options: {
+			replaceWith: '.css-blank'
+		}
+	},
+	'browser': {
+		message: 'css for browser tests',
+	},
+	'browser:replacewith': {
+		message: 'css for browser tests',
+		options: {
+			replaceWith: '.css-blank'
 		}
 	},
 });